正規表現関数切り出し
